# Speedup the Spacemacs startup/loading with load-hints