Hi,

The Spacemacs is heavy and slow on loading (especially for Windows system).
I want share the solution with `load-hints` to speedup the Spacemacs, here are the results, on windows, after enabled the feature, it will reduce the start up time from ~16.7 seconds to ~5.0 seconds; on Ubuntu 20.04, the startup time was reduced from ~1.7s to ~1.4s, 

Windows 11:
342 packages loaded in **16.714s** (e:257 r:19 l:5 b:61)
342 packages loaded in **5.071s** (e:257 r:19 l:5 b:61)
Ubuntu 20.04:
389 packages loaded in **1.729s** (e:302 r:19 l:5 b:63)
389 packages loaded in **1.438s** (e:302 r:19 l:5 b:63)

Just apply this patch and run `spacemacs//package-regenerate-autoloads`, then restart the Spacemacs, check the startup time.